GCP/02_Code/02_CloudRun/01_Api/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
google-cloud-pubsub==2.34.0
pydantic==2.10.3
pydantic_settings==2.6.1